+++ src/templates/mediaproject/src/pages/Root/Root.tsx
@@ -1,0 +1,22 @@
+import { observer } from 'mobx-react';
+import * as React from 'react';
+import { useRoutes } from 'react-router';
+
+import { RouteEnum } from 'config/routes';
+import Home from 'pages/Home';
+import Start from 'pages/Start';
+
+const Root: React.FC = () => {
+  return useRoutes([
+    {
+      path: RouteEnum.home,
+      element: <Home />,
+    },
+    {
+      path: RouteEnum.start,
+      element: <Start />,
+    },
+  ]);
+};
+
+export default observer(Root);