+++ src/templates/mediaproject/src/store/index.ts
@@ -1,0 +1,7 @@
+import endpoints from 'config/endpoints';
+
+import { RootStore } from './RootStore';
+
+const rootStore = new RootStore(endpoints);
+
+export default { rootStore };