+++ src/templates/vk-mini-app/src/components/special/ErrorMessage/ErrorMessage.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
 
-import styles from './ErrorMessage.modules.scss';
+import './ErrorMessage.modules.scss';
 
 const ErrorMessage: React.FC = () => {
   return (
-    <div className={styles.error}>
+    <div styleName="error">
       Произошла ошибка! Перезагрузите приложение или попробуйте позже
     </div>
   );