+++ src/templates/vk-mini-app/src/config/errors.ts
@@ -1,0 +1,7 @@
+export const ErrorsEnum = {
+  badRequest: 'bad_request',
+};
+
+export const errors = {
+  [ErrorsEnum.badRequest]: 'Неверные параметры запроса',
+};