+++ src/templates/vk-mini-app/src/pages/Splash/Splash.tsx
@@ -1,29 +1,16 @@
-import { loadImages } from '@ktsstudio/mediaproject-utils';
 import { observer } from 'mobx-react';
 import * as React from 'react';
 
 import Container from 'components/special/Container';
-import config from 'config/routes';
-import { allStatics } from 'img/config';
-import { useVKHistory } from 'utils/router';
 
-interface Props {
-  onReady(): void;
-}
+import useStartApp from './useStartApp';
+
+type Props = {
+  onReady: VoidFunction;
+};
 
 const Splash: React.FC<Props> = ({ onReady }: Props) => {
-  const { replace } = useVKHistory();
-
-  const [progress, setProgress] = React.useState(0);
-  const imagesLength = allStatics.length;
-  const progressPercent = Math.floor((progress / imagesLength) * 100) || 0;
-
-  React.useEffect(() => {
-    loadImages(allStatics, () => setProgress((p: number) => p + 1)).then(() => {
-      replace({ panel: config.defaultPanel, canSwipeBack: false });
-      onReady();
-    });
-  }, []);
+  const progressPercent = useStartApp(onReady);
 
   return (
     <Container fixedHeight>