+++ src/templates/vk-mini-app/src/store/hooks.ts
@@ -1,8 +1,13 @@
 import { MobXProviderContext } from 'mobx-react';
-import { useContext } from 'react';
+import * as React from 'react';
 
 import { RootStore } from './RootStore';
+import { UserStore } from './UserStore';
 
 export function useStore(): RootStore {
-  return useContext(MobXProviderContext).rootStore as RootStore;
+  return React.useContext(MobXProviderContext).rootStore as RootStore;
+}
+
+export function useUserStore(): UserStore {
+  return useStore().userStore;
 }