+++ src/templates/vk-mini-app/src/store/index.ts
@@ -1,6 +1,9 @@
+import ENDPOINTS from 'config/endpoints';
+import { errors, ErrorsEnum } from 'config/errors';
+
 import { RootStore } from './RootStore';
 
-const rootStore = new RootStore();
+const rootStore = new RootStore(ENDPOINTS, ErrorsEnum, errors);
 
 export default {
   rootStore,